Create README.md
Check your config.ini file before you start the script. 
It wil read the vakues from a erm50 smart meter. In the first run it wil read some daily values and store that into two text files. The second run wil use this value to calculate the daily usage. 
It is a simple script so keep it simple and make a cron job to run it. (do not forget to adjust the location of the config.ini file in the mean script and update it with your path)
I run the script every minute.